move where types are located
